# Add benchmark for updates and get with/without server-side apply