Add a high-level benchmark for repeated updates and get when server-side apply is enabled or not, as a way to compare the impact of enabling the feature.

The output is hard to read though because I haven't managed to disable logs altogether.

Still, the results are:
```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkNoServerSideApply/Get-12                 10460524      10261256      -1.90%
BenchmarkNoServerSideApply/RepeatedUpdates-12     11307965      11474860      +1.48%
BenchmarkServerSideApply/Get-12                   11043359      10495828      -4.96%
BenchmarkServerSideApply/RepeatedUpdates-12       20689213      16360253      -20.92%

benchmark                                         old allocs     new allocs     delta
BenchmarkNoServerSideApply/Get-12                 641            637            -0.62%
BenchmarkNoServerSideApply/RepeatedUpdates-12     2081           2081           +0.00%
BenchmarkServerSideApply/Get-12                   2373           879            -62.96%
BenchmarkServerSideApply/RepeatedUpdates-12       19920          8909           -55.28%
```

Get is not super relevant (especially with/without SSA because it's mostly dominated by overhead), we would have to insert many more objects to actually show the difference.

**What type of PR is this?**
/kind cleanup

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```