@apelisse 2019-08-03
Speed-wise, for PATCH, it went from 1.82x to 1.44x.