@apelisse 2019-08-12
Updated the code, we now get the following results:
```
NoServerSideApply/Get-12               200  34912437 ns/op  3061219 B/op   28242 allocs/op
NoServerSideApply/RepeatedUpdates-12  1000  10976103 ns/op   169518 B/op    2081 allocs/op
NoServerSideApply/Post-12              300  21495704 ns/op  6149313 B/op   56480 allocs/op
ServerSideApply/Get-12                 100  52903182 ns/op  6515623 B/op   45408 allocs/op
ServerSideApply/RepeatedUpdates-12     500  16272308 ns/op   806456 B/op    8913 allocs/op
ServerSideApply/Post-12                300  23594350 ns/op 11460240 B/op  114095 allocs/op
```